modified the Update and delete user.
RequestValidator class: the email, name, and password methods public to be used to validate
Update request as well.

UserController class: modified the update method to take in a custom
request object instead of a User entity and modified the delete user
method to take in a imdbId instead of a user entity.

UserService class: Modified the update method to accept custom request
object, validate the fields and make sure that if a feild is present the
program doesn't crash or replace the data with the empty field. Modified
the delete method to take in a imdbId string instead of a user entitiy.

+++ src/main/java/movieApi/movies/service/UserService.java
@@ -3,10 +3,12 @@ package movieApi.movies.service;
 import com.mongodb.client.result.UpdateResult;
 import movieApi.movies.converter.Converter;
 import movieApi.movies.dto.request.CreateUserRequest;
+import movieApi.movies.dto.request.UpdateUserRequest;
 import movieApi.movies.dto.response.PrivateUserDTO;
 import movieApi.movies.dto.response.PublicUserDTO;
 import movieApi.movies.entity.User;
 import movieApi.movies.exception.InvalidHTTPRequestException;
+import movieApi.movies.exception.UserNotFoundException;
 import movieApi.movies.repository.UserRepository;
 import movieApi.movies.utils.CustomIdMaker;
 import movieApi.movies.utils.RequestValidator;
@@ -44,15 +46,15 @@ public class UserService {
 
     public PublicUserDTO getPublicUserByUsername(String username) {
         User user = template.findOne(Query.query(
-                                Criteria.where("username").is(username)),
-                                User.class);
+                        Criteria.where("username").is(username)) ,
+                User.class);
         assert user != null;
         return Converter.userToPublicDTO(user);
     }
 
     public PublicUserDTO getPublicUserByEmail(String email) {
         User user = template.findOne(Query.query(
-                        Criteria.where("email").is(email)),
+                        Criteria.where("email").is(email)) ,
                 User.class);
         assert user != null;
         return Converter.userToPublicDTO(user);
@@ -77,37 +79,59 @@ public class UserService {
             }
         }
 
-        User createdUser = repo.insert(new User(imdbId,
-                user.firstName(),
-                user.lastName(),
-                user.username(),
-                user.password(),
-                user.email(),
+        User createdUser = repo.insert(new User(imdbId ,
+                user.firstName() ,
+                user.lastName() ,
+                user.username() ,
+                user.password() ,
+                user.email() ,
                 new ArrayList<>()));
 
         return Converter.userToPrivateDTO(createdUser);
     }
 
-    public PrivateUserDTO updateExistingUser(User user) {
-        Update update = Update
-                .update("firstName", user.getFirstName())
-                .set("lastName", user.getLastName())
-                .set("username", user.getUsername())
-                .set("password", user.getPassword())
-                .set("email", user.getEmail());
+    public PrivateUserDTO updateExistingUser(UpdateUserRequest request) throws InvalidHTTPRequestException {
+        if (request == null) throw new InvalidHTTPRequestException("Request is null");
 
-        Query query = new Query(Criteria.where("imdbId").is(user.getImdbId()));
+        String imdbId = request.imdbId(),
+                firstName = request.firstName(),
+                lastName = request.lastName(),
+                username = request.username(),
+                password = request.password(),
+                email = request.email();
 
-        UpdateResult result = template.updateFirst(query, update, User.class);
+
+        validator.isValidName(firstName);
+        validator.isValidName(lastName);
+        validator.isValidPassword(password);
+        validator.isValidEmail(email);
+        validator.doesUserExist(username , email);
+
+        Update update = new Update();
+
+        if (!firstName.isEmpty()) update.set("firstName" , firstName);
+        if (!lastName.isEmpty()) update.set("lastName" , lastName);
+        if (!username.isEmpty()) update.set("username" , username);
+        if (!password.isEmpty()) update.set("password" , password);
+        if (!email.isEmpty()) update.set("email" , email);
+
+        Query query = new Query(Criteria.where("imdbId").is(imdbId));
+
+        UpdateResult result = template.updateFirst(query , update , User.class);
 
         if (result.getModifiedCount() > 0) {
-            return Converter.userToPrivateDTO(user);
+            return Converter.userToPrivateDTO(repo.findUserByImdbId(imdbId)
+                    .orElseThrow(UserNotFoundException::new));
         } else {
             throw new RuntimeException("internal server error at service.");
         }
     }
 
-    public void deleteUserFromDB(User user) {
+    public void deleteUserFromDB(String imdbId) throws InvalidHTTPRequestException {
+        if (imdbId.isEmpty()) throw new InvalidHTTPRequestException("request null");
+
+        User user = repo.findUserByImdbId(imdbId).orElseThrow(UserNotFoundException::new);
+
         repo.delete(user);
     }
 }
